Fix median calculation when n=1

+++ courses/views.py
@@ -174,9 +174,9 @@ def median(pool):
     copy = sorted(pool)
     size = len(copy)
     if size % 2 == 1:
-        return copy[(size - 1) / 2]
+        return copy[int((size - 1) / 2)]
     else:
-        return (copy[size/2 - 1] + copy[size/2]) / 2
+        return (copy[int(size/2 - 1)] + copy[int(size/2)]) / 2
 
 
 def grades_csv(course):
